+++ mol_assembler.py
@@ -11,8 +11,8 @@ RDLogger.DisableLog('rdApp.*')  # Hides ALL warnings, errors, and info from RDKi
 from rdkit import Chem
 from rdkit.Chem import Draw
 
-from tools.data_utils import log_event, _open_dialogs
-from tools.data_utils import (
+from data_utils import log_event, _open_dialogs
+from data_utils import (
     results_dir, mineral_favs_path, element_favs_path, ptable_path,
     mineral_db_path, gallery_dir, gallery_meta_path, log_path, chain_log_path,
     exports_dir, settings_path, library_file, load_settings, load_element_data, ai_chatlogs_dir